Correcting For Get_Set Methods
This commit should address getting the EKF origin automatically from the autopilot upon first detection of the heartbeat. The current release of Arducopter will publish both EKF and Home when the Home position is requested. Within this commit there is also handling of message_request however is not currently supported within the 3.6 branch of arducopter.